--- data/subgen_gene_plots/antibiotic_response_genes.txt
@@ -1,106 +1,0 @@
-mdtQ
-marA
-arnB
-ralR
-gyrA
-tufB
-rlmN
-priA
-eptA
-mdtG
-rpsL
-hipA
-basR
-rplF
-macA
-mdtL
-emrB
-ampE
-tehA
-arnC
-rcbA
-gyrB
-mdtO
-rpoB
-dacB
-mrdA
-pbpC
-rsmA
-rcsB
-rplD
-thrS
-tehB
-ampC
-acrZ
-rpsD
-rplV
-lpxD
-acrE
-marR
-rpsQ
-bcr
-parE
-rpoC
-tufA
-ftsE
-fabI
-yjjX
-folA
-pmrD
-marB
-rpsE
-ileS
-ldtD
-mrcB
-ymdB
-aroK
-acrD
-acrB
-ybhG
-yqaE
-rpmG
-nimT
-prc
-tolC
-sbmA
-pyrR
-kilR
-sseA
-mdfA
-bamE
-mprA
-relE
-emrY
-emrK
-ycjU
-ybcM
-mlaB
-ykfM
-entS
-rfbC
-rfbX
-yibA
-efeO
-yadC
-alaA
-arnD
-yafQ
-arnA
-macB
-ybjG
-mdtH
-bacA
-blr
-fsr
-mdtM
-mdtF
-mdtE
-mdtK
-yojI
-mdtN
-mdtP
-emrA
-acrA
-mrcA
-ompF
-ompC
